pyproject.toml
[project]
name = "kd3005p"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyserial>=3.5",
]
